refactor: extract class PatchFactoryGpu
